templates, layers, requestbearer reworked

+++ api/user/delete.js
@@ -1,31 +1,33 @@
-const requestBearer = require('../../mod/requestBearer');
-
-const acl = require('../../mod/auth/acl')();
-
-const mailer = require('../../mod/mailer');
-
-module.exports = (req, res) => requestBearer(req, res, [ handler ], {
+const auth = require('../../mod/auth/handler')({
   admin_user: true
-});
+})
 
-async function handler(req, res) {
+const acl = require('../../mod/auth/acl')()
 
-  const email = req.query.email.replace(/\s+/g, '');
+const mailer = require('../../mod/mailer')
+
+module.exports = async (req, res) => {
+
+  await auth(req, res)
+
+  if (res.finished) return
+
+  const email = req.query.email.replace(/\s+/g, '')
 
   // Delete user account in ACL.
   var rows = await acl(`
     DELETE FROM acl_schema.acl_table
-    WHERE lower(email) = lower($1);`, [email]);
+    WHERE lower(email) = lower($1);`, [email])
 
-  if (rows instanceof Error) return res.status(500).send('Failed to query PostGIS table.');
+  if (rows instanceof Error) return res.status(500).send('Failed to query PostGIS table.')
 
   // Sent email to inform user that their account has been deleted.
   await mailer({
     to: email,
     subject: `This ${req.headers.host}${process.env.DIR || ''} account has been deleted.`,
     text: `You will no longer be able to log in to ${req.headers.host}${process.env.DIR || ''}`
-  });
+  })
 
-  res.send('User account deleted.');
+  res.send('User account deleted.')
 
-};
+}
